Extract shared model-test factory helpers to test/support
Moves the factory-equivalent builders (create_user/product/purchase/
subscription/workflow/variant/installment/etc.) out of
installment_test.rb into a shared ModelFactories module, mixed into every
ActiveSupport::TestCase. installment_test keeps only its installment-
specific context helpers.

This is groundwork for the next model-spec ports (subscription, purchase,
link, all #5801): they build the same objects, so sharing one
implementation keeps those ports mechanical instead of re-deriving each
factory. No behavior change — installment_test and the full test/ tree
stay green (112 / 142 runs).
