+++ test/support/model_factories.rb
@@ -1,0 +1,302 @@
+# frozen_string_literal: true
+
+# Factory-equivalent builders for the Minitest + fixtures suite (#5801).
+#
+# The suite has no FactoryBot, but complex hub models (Installment, Purchase,
+# Subscription, Link, …) are genuinely per-test and can't be static fixtures.
+# These helpers build them with real `Model.create!` — callbacks run, so the
+# objects behave like factory-built ones — and mirror the essential attributes
+# each factory sets. They're mixed into every ActiveSupport::TestCase so model
+# ports can share one implementation.
+#
+# Naming mirrors the factories (create_user, create_purchase, …) to keep the
+# port mechanical. Add new builders here rather than re-deriving them per file.
+module ModelFactories
+  # Route URL helpers, without mixing them into the test class directly:
+  # `include Rails.application.routes.url_helpers` defines helpers like
+  # `test_ping_url`, and Minitest runs every /^test_/ method as a test. Call as
+  # `routes.some_url(...)`.
+  def routes
+    Rails.application.routes.url_helpers
+  end
+
+  def unique_suffix
+    SecureRandom.hex(8)
+  end
+
+  # A distinct IP per purchase within a test (see create_purchase): not_double_charged
+  # keys on email + ip_address + link, so a shared/nil IP makes two sales to the
+  # same buyer look like a duplicate charge. The factory uses a random IP.
+  def unique_ip
+    @ip_seq ||= 0
+    @ip_seq += 1
+    "10.0.#{@ip_seq / 254}.#{@ip_seq % 254 + 1}"
+  end
+
+  # Deactivate then restart a subscription, recording the events code paths like
+  # Installment#delivery_due? read to reason about post-resubscribe timing.
+  def resubscribe(subscription)
+    subscription.update!(deactivated_at: nil)
+    SubscriptionEvent.create!(subscription:, event_type: :deactivated, occurred_at: 10.days.ago)
+    SubscriptionEvent.create!(subscription:, event_type: :restarted, occurred_at: 2.days.ago)
+  end
+
+  def moderation_result(passed:, reasons: [])
+    ContentModeration::ModerateRecordService::CheckResult.new(passed:, reasons:)
+  end
+
+  def create_user(**attrs)
+    User.create!({
+      email: "user-#{unique_suffix}@example.com",
+      username: "u#{unique_suffix}",
+      password: "test-password-123!",
+      confirmed_at: Time.current,
+      user_risk_state: "not_reviewed",
+    }.merge(attrs))
+  end
+
+  def create_product(user: nil, **attrs)
+    Link.create!({ user: user || create_user, name: "The Works of Edgar Gumstein", price_cents: 100 }.merge(attrs))
+  end
+
+  # A recurring (non-tiered) membership product, mirroring :subscription_product.
+  def create_subscription_product(user: nil, **attrs)
+    Link.create!({
+      user: user || create_user,
+      name: "Membership",
+      price_cents: 100,
+      is_recurring_billing: true,
+      subscription_duration: "monthly",
+      is_tiered_membership: false,
+    }.merge(attrs))
+  end
+
+  def create_variant_category(link: nil, **attrs)
+    VariantCategory.create!({ link: link || create_product, title: "Category" }.merge(attrs))
+  end
+
+  def create_variant(variant_category: nil, **attrs)
+    Variant.create!({ variant_category: variant_category || create_variant_category, name: "Untitled" }.merge(attrs))
+  end
+
+  # Mirrors the :installment factories. Product/variant posts hang off a product;
+  # seller/follower/affiliate/audience posts hang off a seller with no link.
+  def build_installment(installment_type: "product", link: :default, seller: :default, base_variant: nil, **attrs)
+    case installment_type
+    when "product"
+      link = create_product if link == :default
+      seller = link&.user if seller == :default
+    when "variant"
+      base_variant ||= create_variant
+      link = base_variant.link if link == :default
+      seller = link&.user if seller == :default
+    else # seller, follower, affiliate, audience, abandoned_cart
+      link = nil if link == :default
+      seller = create_user if seller == :default
+    end
+    Installment.new({
+      installment_type:,
+      link:,
+      seller:,
+      base_variant:,
+      message: "<p>A message.</p>",
+      name: "A Post",
+      send_emails: true,
+      shown_on_profile: false,
+      allow_comments: true,
+    }.merge(attrs))
+  end
+
+  def create_installment(**args)
+    build_installment(**args).tap(&:save!)
+  end
+
+  # Mirrors the base :purchase factory (see test/fixtures/purchases.yml for the
+  # same money math): a successful sale on the platform Stripe account.
+  # calculate_fees is an after(:build) hook in the factory, so invoke it
+  # explicitly before saving.
+  def create_purchase(link:, seller: :default, purchaser: nil, variant_attributes: nil, **attrs)
+    seller = link.user if seller == :default
+    price_cents = attrs.delete(:price_cents) || link.price_cents || 100
+    purchase = Purchase.new({
+      link:,
+      seller:,
+      price_cents:,
+      total_transaction_cents: price_cents,
+      displayed_price_cents: price_cents,
+      ip_address: unique_ip,
+      email: "buyer-#{unique_suffix}@example.com",
+      purchase_state: "successful",
+      succeeded_at: Time.current,
+      card_type: "visa",
+      card_visual: "**** **** **** 4062",
+      card_country: "US",
+      stripe_fingerprint: "shfbeg5142fff",
+      stripe_transaction_id: "2763276372637263",
+      charge_processor_id: "stripe",
+      merchant_account: merchant_accounts(:gumroad_stripe),
+    }.merge(attrs))
+    purchase.purchaser = purchaser if purchaser
+    purchase.variant_attributes = variant_attributes if variant_attributes
+    purchase.send(:calculate_fees)
+    purchase.save!
+    purchase
+  end
+
+  # A $0 sale (mirrors :free_purchase): no charge processor, card, or stripe ids.
+  def create_free_purchase(link:, seller: :default, **attrs)
+    create_purchase(
+      link:, seller:, price_cents: 0,
+      charge_processor_id: nil, merchant_account: nil,
+      card_type: nil, card_visual: nil,
+      stripe_fingerprint: nil, stripe_transaction_id: nil,
+      **attrs
+    )
+  end
+
+  # An original subscription sale on a plain (non-tiered) recurring product —
+  # enough for code that only cares about the subscription, not membership tiers.
+  def create_membership_purchase(link: nil, created_at: nil, **attrs)
+    link ||= create_subscription_product
+    subscription = create_subscription(link:)
+    create_purchase(link:, subscription:, is_original_subscription_purchase: true, created_at:, **attrs)
+  end
+
+  # Subscriptions must have a payment_option before they validate, so build it
+  # (priced at the product's default price) before saving, like the :subscription
+  # factory's before(:create) hook does.
+  def create_subscription(link: nil, user: nil, **attrs)
+    link ||= create_product
+    user ||= create_user
+    subscription = Subscription.new({ link:, user:, is_installment_plan: false }.merge(attrs))
+    subscription.payment_options << PaymentOption.new(subscription:, price: link.default_price)
+    subscription.save!
+    subscription
+  end
+
+  # Mirrors the :workflow factories. Product/variant workflows hang off a
+  # product; the rest have no link.
+  def create_workflow(workflow_type: "product", seller: nil, link: :default, **attrs)
+    seller ||= create_user
+    if %w[product variant].include?(workflow_type)
+      link = create_product(user: seller) if link == :default
+    elsif link == :default
+      link = nil
+    end
+    Workflow.create!({ seller:, link:, name: "my workflow", workflow_type:, workflow_trigger: nil }.merge(attrs))
+  end
+
+  # A published installment attached to a workflow, with an installment_rule
+  # (mirrors :workflow_installment). delayed_delivery_time drives delivery_due?.
+  def create_workflow_installment(workflow:, link: nil, published_at: nil, delayed_delivery_time: 0, **attrs)
+    installment = Installment.create!({
+      workflow:,
+      seller: workflow.seller,
+      link: link || workflow.link,
+      installment_type: workflow.workflow_type,
+      base_variant: workflow.base_variant,
+      send_emails: true,
+      message: "<p>A message.</p>",
+      name: "A Post",
+      published_at:,
+      workflow_installment_published_once_already: published_at.present?,
+    }.merge(attrs))
+    InstallmentRule.create!(installment:, delayed_delivery_time:, to_be_published_at: 1.week.from_now, time_period: "hour")
+    installment
+  end
+
+  # An abandoned-cart workflow with its seeded installment (mirrors the
+  # abandoned_cart_workflow factory's after(:create) hook).
+  def create_abandoned_cart_workflow(seller:, **attrs)
+    workflow = Workflow.create!({ seller:, link: nil, name: "my workflow", workflow_type: "abandoned_cart", workflow_trigger: nil }.merge(attrs))
+    Installment.create!(
+      workflow:,
+      seller: workflow.seller,
+      link: workflow.link,
+      installment_type: workflow.workflow_type,
+      base_variant: workflow.base_variant,
+      send_emails: true,
+      published_at: workflow.published_at,
+      workflow_installment_published_once_already: workflow.published_at.present?,
+      name: "You left something in your cart",
+      message: "When you're ready to buy, complete checking out.<product-list-placeholder />Thanks!"
+    )
+    workflow
+  end
+
+  def create_offer_code(user:, products:, amount_cents: 100, **attrs)
+    offer_code = OfferCode.new({ user:, products:, code: "off#{unique_suffix[0, 8]}", amount_cents:, currency_type: user.currency_type }.merge(attrs))
+    offer_code.user_id = products.first.user_id if products.present? # mirrors the factory's before(:create)
+    offer_code.save!
+    offer_code
+  end
+
+  def create_upsell(seller:, product:, offer_code: nil, **attrs)
+    Upsell.create!({
+      name: "Upsell",
+      seller:,
+      product:,
+      text: "Take advantage of this excellent offer!",
+      description: "This offer will only last for a few weeks.",
+      cross_sell: false,
+      offer_code:,
+    }.merge(attrs))
+  end
+
+  def create_affiliate_user(**attrs)
+    create_user(**attrs)
+  end
+
+  def create_direct_affiliate(seller:, affiliate_user: nil, **attrs)
+    DirectAffiliate.create!({
+      affiliate_user: affiliate_user || create_affiliate_user,
+      seller:,
+      affiliate_basis_points: 300,
+      send_posts: true,
+    }.merge(attrs))
+  end
+
+  def create_blast(post:, **attrs)
+    PostEmailBlast.create!({
+      post:,
+      seller: post.seller,
+      requested_at: 30.minutes.ago,
+      started_at: 25.minutes.ago,
+      first_email_delivered_at: 20.minutes.ago,
+      last_email_delivered_at: 10.minutes.ago,
+      delivery_count: 1500,
+    }.merge(attrs))
+  end
+
+  def create_product_file(installment: nil, link: nil, **attrs)
+    ProductFile.create!({
+      url: "#{S3_BASE_URL}specs/#{unique_suffix}.pdf",
+      filetype: "pdf",
+      filegroup: "document",
+      installment:,
+      link:,
+    }.merge(attrs))
+  end
+
+  PRODUCT_FILE_SHAPES = {
+    streamable_video: { url: "#{S3_BASE_URL}specs/ScreenRecording.mov", filetype: "mov", filegroup: "video" },
+    readable_document: { url: "#{S3_BASE_URL}specs/billion-dollar-company-chapter-0.pdf", filetype: "pdf", filegroup: "document" },
+  }.freeze
+
+  def create_installment_file(shape, **attrs)
+    ProductFile.create!(PRODUCT_FILE_SHAPES.fetch(shape).merge(attrs))
+  end
+
+  # CreatorContactingCustomersEmailInfo rows, mirroring the nested
+  # creator_contacting_customers_email_info_{sent,delivered,opened} factories:
+  # each state carries the timestamps of the states it descends from.
+  def create_email_info(installment:, purchase:, state: "sent", email_name: "purchase_installment")
+    attrs = { installment:, purchase:, email_name:, state: }
+    attrs[:sent_at] = Time.current if %w[sent delivered opened].include?(state)
+    attrs[:delivered_at] = Time.current if %w[delivered opened].include?(state)
+    attrs[:opened_at] = Time.current if state == "opened"
+    CreatorContactingCustomersEmailInfo.create!(attrs)
+  end
+end
+
+ActiveSupport::TestCase.include(ModelFactories)
